работает регистрация - v.1.3

+++ src/main/java/ru/coolteam/earnpocketmoney/repositories/RoleRepository.java
@@ -5,10 +5,10 @@ import org.springframework.stereotype.Repository;
 import ru.coolteam.earnpocketmoney.models.Role;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+//public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 
     Role findByRole(String role);
-
 
 }
